backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # --- Backend ---
    "fastapi",
    "uvicorn[standard]",
    # --- Frontend ---
    "requests",
    "httpx>=0.27.0",
    # --- RAG pipeline ---
    "langchain",
    "langchain-community",
    "langchain-text-splitters",
    "langchain-cohere>=0.2.0",
    "langchain-openai>=0.1.7",
    "weaviate-client",
    "sentence-transformers",
    # --- Document loaders ---
    "pypdf",
    "python-docx",
    "markdown",
    # --- LLM provider ---
    "openai>=1.30.0",
    # --- Utility ---
    "tiktoken", # optional, helps with token counting
    "langchain-huggingface>=0.3.1",
    "langchain-weaviate>=0.0.1,<1.0.0",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "supabase>=2.19.0",
    "groq>=0.11.0",
    "httptools>=0.6.4",
    "uvloop>=0.21.0",
    "elevenlabs>=2.16.0",
    "pydantic-settings>=2.11.0",
    "azure-cognitiveservices-speech>=1.46.0",
    "ipykernel>=7.0.1",
    # --- Observability ---
    "opentelemetry-sdk>=1.23.0",
    "opentelemetry-exporter-otlp-proto-http>=1.23.0",
    "openinference-instrumentation>=0.1.41",
    "openinference-instrumentation-openai>=0.1.39",
    "openinference-instrumentation-langchain>=0.1.54",
    "opentelemetry-instrumentation-fastapi>=0.59b0",
    "redis>=7.0.1",
    "fakeredis>=2.32.1",
    "prometheus-fastapi-instrumentator>=6.1.1",
    "prometheus_client>=0.20.0",
]

[tool.uv.workspace]
members = [
    "MCP",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "integration: integration tests that require running services (e.g., MCP server)",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=0.23",
]
